made whole product clickable (by doing so, did not had to add eslint exceptions), exported styles and a function for modal placement
