+++ front-end/src/components/productTable/ProductItem.js
@@ -1,6 +1,3 @@
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* upper two are required for the eslint in <b onClick> */
 import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
@@ -12,14 +9,14 @@ function ProductItem(props) {
   const { classes, product, selectProduct } = props;
   return (
     <Grid item>
-      <Paper className={classes.paper} component="div">
+      <Paper className={classes.paper} component="div" onClick={selectProduct}>
         <img
           src={`data:image/png;base64,${product.image}`}
           alt={product.title}
           className={classes.image}
         />
         <Divider />
-        <b onClick={selectProduct}>
+        <b>
           {product.title}
           <span className={classes.price}>{product.price}€</span>
         </b>
